+++ client/group.js
@@ -8,8 +8,8 @@ Group = function(router, options, parent) {
 
   this._router = router;
   this.prefix = options.prefix || '';
+  this.name = options.name;
 
-  this._middlewares = options.middlewares || [];
   this._triggersEnter = options.triggersEnter || [];
   this._triggersExit = options.triggersExit || [];
   this._subscriptions = options.subscriptions || Function.prototype;
@@ -17,26 +17,22 @@ Group = function(router, options, parent) {
   this.parent = parent;
   if (this.parent) {
     this.prefix = parent.prefix + this.prefix;
-    this._middlewares = parent._middlewares.concat(this._middlewares);
 
     this._triggersEnter = parent._triggersEnter.concat(this._triggersEnter);
     this._triggersExit = this._triggersExit.concat(parent._triggersExit);
   }
 };
 
-Group.prototype.route = function(path, options, group) {
+Group.prototype.route = function(pathDef, options, group) {
   options = options || {};
 
-  if (!/^\/.*/.test(path)) {
+  if (!/^\/.*/.test(pathDef)) {
     var message = "route's path must start with '/'";
     throw new Error(message);
   }
 
   group = group || this;
-  path = this.prefix + path;
-
-  var middlewares = options.middlewares || [];
-  options.middlewares = this._middlewares.concat(middlewares);
+  pathDef = this.prefix + pathDef;
 
   var triggersEnter = options.triggersEnter || [];
   options.triggersEnter = this._triggersEnter.concat(triggersEnter);
@@ -44,7 +40,7 @@ Group.prototype.route = function(path, options, group) {
   var triggersExit = options.triggersExit || [];
   options.triggersExit = triggersExit.concat(this._triggersExit);
 
-  return this._router.route(path, options, group);
+  return this._router.route(pathDef, options, group);
 };
 
 Group.prototype.group = function(options) {