+++ client/route.js
@@ -1,7 +1,12 @@
-Route = function(router, path, options, group) {
+Route = function(router, pathDef, options, group) {
   options = options || {};
 
-  this.path = path;
+  this.options = options;
+  this.pathDef = pathDef
+
+  // Route.path is deprecated and will be removed in 3.0
+  this.path = pathDef;
+
   if (options.name) {
     this.name = options.name;
   }
@@ -10,7 +15,6 @@ Route = function(router, path, options, group) {
   this._subscriptions = options.subscriptions || Function.prototype;
   this._triggersEnter = options.triggersEnter || [];
   this._triggersExit = options.triggersExit || [];
-  this._middlewares = options.middlewares || [];
   this._subsMap = {};
   this._router = router;
 
@@ -42,45 +46,9 @@ Route.prototype.getAllSubscriptions = function() {
   return this._subsMap;
 };
 
-
-Route.prototype._processTriggersEnter = function(context) {
-  _.each(this._triggersEnter, function(fn) {
-    if (typeof fn === 'function') {
-      fn(context);
-    }
-  });
-};
-
-Route.prototype._processMiddlewares = function(context, after) {
-  var currentIndex = 0;
-  var self = this;
-
-  runMiddleware();
-  function runMiddleware() {
-    var fn = self._middlewares[currentIndex++];
-    if(fn) {
-      console.warn("'middleware' is deprecated. Use 'triggers' instead");
-      fn(context.path, function(redirectPath) {
-        if(redirectPath) {
-          return self._router.redirect(redirectPath);
-        } else {
-          runMiddleware();
-        }
-      });
-    } else {
-      after();
-    }
-  }
-};
-
 Route.prototype.callAction = function(current) {
   var self = this;
-
-  self._processTriggersEnter(current);
-
-  self._processMiddlewares(current.context, function() {
-    self._action(current.params, current.queryParams);
-  });
+  self._action(current.params, current.queryParams);
 };
 
 Route.prototype.callSubscriptions = function(current) {