+++ client/router.js
@@ -1,4 +1,5 @@
 Router = function () {
+  var self = this;
   this.globals = [];
   this.subscriptions = Function.prototype;
 
@@ -10,58 +11,114 @@ Router = function () {
 
   this._globalRoute = new Route(this);
 
+  // holds onRoute callbacks
+  this._onRouteCallbacks = [];
+
+  // if _askedToWait is true. We don't automatically start the router
+  // in Meteor.startup callback. (see client/_init.js)
+  // Instead user need to call `.initialize()
+  this._askedToWait = false;
+  this._initialized = false;
   this._triggersEnter = [];
   this._triggersExit = [];
-  this._middleware = [];
   this._routes = [];
   this._routesMap = {};
   this._updateCallbacks();
   this.notFound = this.notfound = null;
   // indicate it's okay (or not okay) to run the tracker
   // when doing subscriptions
-  // using a number and increment it help us to support FlowRouter.go() 
+  // using a number and increment it help us to support FlowRouter.go()
   // and legitimate reruns inside tracker on the same event loop.
   // this is a solution for #145
   this.safeToRun = 0;
 
-  var self = this;
-  this.triggers = {
-    enter: self._getRegisterTriggersFn(self._triggersEnter),
-    exit: self._getRegisterTriggersFn(self._triggersExit)
-  };
+  // Meteor exposes to the client the path prefix that was defined using the
+  // ROOT_URL environement variable on the server using the global runtime
+  // configuration. See #315.
+  this._basePath = __meteor_runtime_config__.ROOT_URL_PATH_PREFIX || '';
+
+  // this is a chain contains a list of old routes
+  // most of the time, there is only one old route
+  // but when it's the time for a trigger redirect we've a chain
+  this._oldRouteChain = [];
 
   this.env = {
     replaceState: new Meteor.EnvironmentVariable(),
-    reload: new Meteor.EnvironmentVariable()
+    reload: new Meteor.EnvironmentVariable(),
+    trailingSlash: new Meteor.EnvironmentVariable()
   };
+
+  // redirect function used inside triggers
+  this._redirectFn = function(pathDef, fields, queryParams) {
+    if (/^http(s)?:\/\//.test(pathDef)) {
+        var message = "Redirects to URLs outside of the app are not supported in this version of Flow Router. Use 'window.location = yourUrl' instead";
+        throw new Error(message);
+    }
+    self.withReplaceState(function() {
+      var path = FlowRouter.path(pathDef, fields, queryParams);
+      self._page.redirect(path);
+    });
+  };
+  this._initTriggersAPI();
 };
 
-Router.prototype.route = function(path, options, group) {
-  if (!/^\/.*/.test(path)) {
+Router.prototype.route = function(pathDef, options, group) {
+  if (!/^\/.*/.test(pathDef)) {
     var message = "route's path must start with '/'";
     throw new Error(message);
   }
 
   options = options || {};
   var self = this;
-  var route = new Route(this, path, options, group);
+  var route = new Route(this, pathDef, options, group);
 
-  route._handler = function (context, next) {
+  // calls when the page route being activates
+  route._actionHandle = function (context, next) {
     var oldRoute = self._current.route;
+    self._oldRouteChain.push(oldRoute);
+
+    var queryParams = self._qs.parse(context.querystring);
+    // _qs.parse() gives us a object without prototypes,
+    // created with Object.create(null)
+    // Meteor's check doesn't play nice with it.
+    // So, we need to fix it by cloning it.
+    // see more: https://github.com/meteorhacks/flow-router/issues/164
+    queryParams = JSON.parse(JSON.stringify(queryParams));
 
     self._current = {
       path: context.path,
       context: context,
       params: context.params,
-      queryParams: self._qs.parse(context.querystring),
+      queryParams: queryParams,
       route: route,
       oldRoute: oldRoute
     };
 
-    // to backward compatibility
-    self._current.params.query = self._current.queryParams;
+    // we need to invalidate if all the triggers have been completed
+    // if not that means, we've been redirected to another path
+    // then we don't need to invalidate
+    var afterAllTriggersRan = function() {
+      self._invalidateTracker();
+    };
 
-    self._invalidateTracker();
+    var triggers = self._triggersEnter.concat(route._triggersEnter);
+    Triggers.runTriggers(
+      triggers,
+      self._current,
+      self._redirectFn,
+      afterAllTriggersRan
+    );
+  };
+
+  // calls when you exit from the page js route
+  route._exitHandle = function(context, next) {
+    var triggers = self._triggersExit.concat(route._triggersExit);
+    Triggers.runTriggers(
+      triggers,
+      self._current,
+      self._redirectFn,
+      next
+    );
   };
 
   this._routes.push(route);
@@ -70,6 +127,7 @@ Router.prototype.route = function(path, options, group) {
   }
 
   this._updateCallbacks();
+  this._triggerRouteRegister(route);
 
   return route;
 };
@@ -80,24 +138,43 @@ Router.prototype.group = function(options) {
 
 Router.prototype.path = function(pathDef, fields, queryParams) {
   if (this._routesMap[pathDef]) {
-    pathDef = this._routesMap[pathDef].path;
+    pathDef = this._routesMap[pathDef].pathDef;
   }
 
-  // remove trailing slash(es)
-  // but keep the root slash if it's the only one
-  pathDef = pathDef.match(/^\/{1}$/) ? pathDef : pathDef.replace(/\/+$/, "");
+  var path = "";
+
+  // Prefix the path with the router global prefix
+  if (this._basePath) {
+    path += "/" + this._basePath + "/";
+  }
 
   fields = fields || {};
   var regExp = /(:[\w\(\)\\\+\*\.\?]+)+/g;
-  var path = pathDef.replace(regExp, function(key) {
+  path += pathDef.replace(regExp, function(key) {
     var firstRegexpChar = key.indexOf("(");
     // get the content behind : and (\\d+/)
     key = key.substring(1, (firstRegexpChar > 0)? firstRegexpChar: undefined);
     // remove +?*
     key = key.replace(/[\+\*\?]+/g, "");
 
-    return encodeURIComponent(fields[key] || "");
+    // this is to allow page js to keep the custom characters as it is
+    // we need to encode 2 times otherwise "/" char does not work properly
+    // So, in that case, when I includes "/" it will think it's a part of the
+    // route. encoding 2times fixes it
+    return encodeURIComponent(encodeURIComponent(fields[key] || ""));
   });
+
+  // Replace multiple slashes with single slash
+  path = path.replace(/\/\/+/g, "/");
+
+  // remove trailing slash
+  // but keep the root slash if it's the only one
+  path = path.match(/^\/{1}$/) ? path: path.replace(/\/$/, "");
+
+  // explictly asked to add a trailing slash
+  if(this.env.trailingSlash.get() && _.last(path) !== "/") {
+    path += "/";
+  }
 
   var strQueryParams = this._qs.stringify(queryParams || {});
   if(strQueryParams) {
@@ -109,10 +186,10 @@ Router.prototype.path = function(pathDef, fields, queryParams) {
 
 Router.prototype.go = function(pathDef, fields, queryParams) {
   var path = this.path(pathDef, fields, queryParams);
-  
+
   var useReplaceState = this.env.replaceState.get();
   if(useReplaceState) {
-    this._page.redirect(path);
+    this._page.replace(path);
   } else {
     this._page(path);
   }
@@ -133,7 +210,7 @@ Router.prototype.redirect = function(path) {
 Router.prototype.setParams = function(newParams) {
   if(!this._current.route) {return false;}
 
-  var pathDef = this._current.route.path;
+  var pathDef = this._current.route.pathDef;
   var existingParams = this._current.params;
   var params = {};
   _.each(_.keys(existingParams), function(key) {
@@ -159,7 +236,7 @@ Router.prototype.setQueryParams = function(newParams) {
     }
   }
 
-  var pathDef = this._current.route.path;
+  var pathDef = this._current.route.pathDef;
   var params = this._current.params;
   this.go(pathDef, params, queryParams);
   return true;
@@ -169,22 +246,18 @@ Router.prototype.setQueryParams = function(newParams) {
 // This is by design. use .getParam() instead
 // If you really need to watch the path change, use .watchPathChange()
 Router.prototype.current = function() {
-  return this._current;
-};
-
-Router.prototype.reactiveCurrent = function() {
-  var warnMessage = 
-    ".reactiveCurrent() is deprecated. " +
-    "Use .watchPathChange() instead";
-  console.warn(warnMessage);
-  
-  this.watchPathChange();
-  return this.current();
+  // We can't trust outside, that's why we clone this
+  // Anyway, we can't clone the whole object since it has non-jsonable values
+  // That's why we clone what's really needed.
+  var current = _.clone(this._current);
+  current.queryParams = EJSON.clone(current.queryParams);
+  current.params = EJSON.clone(current.params);
+  return current;
 };
 
 // Implementing Reactive APIs
 var reactiveApis = [
-  'getParam', 'getQueryParam', 
+  'getParam', 'getQueryParam',
   'getRouteName', 'watchPathChange'
 ];
 reactiveApis.forEach(function(api) {
@@ -202,35 +275,6 @@ reactiveApis.forEach(function(api) {
     return currentRoute[api].call(currentRoute, arg1);
   };
 });
-
-Router.prototype.middleware = function(middlewareFn) {
-  console.warn("'middleware' is deprecated. Use 'triggers' instead");
-  var self = this;
-  var mw = function(ctx, next) {
-    // make sure middlewars run after Meteor has been initialized
-    // this is very important for specially for fast render and Meteor.user()
-    // availability
-    Meteor.startup(function() {
-      middlewareFn(ctx.pathname, processNext);
-    });
-
-    function processNext(path) {
-      if(path) {
-        return self._page.redirect(path);
-      }
-      next();
-    }
-  };
-
-  this._middleware.push(mw);
-  this._updateCallbacks();
-  return this;
-};
-
-Router.prototype.ready = function() {
-  console.warn("'FlowRouter.ready()' is deprecated. Use 'FlowRouter.subsReady()' instead");
-  return this.subsReady.apply(this, arguments);
-};
 
 Router.prototype.subsReady = function() {
   var callback = null;
@@ -285,6 +329,10 @@ Router.prototype.withReplaceState = function(fn) {
   return this.env.replaceState.withValue(true, fn);
 };
 
+Router.prototype.withTrailingSlash = function(fn) {
+  return this.env.trailingSlash.withValue(true, fn);
+};
+
 Router.prototype._notfoundRoute = function(context) {
   this._current = {
     path: context.path,
@@ -304,7 +352,13 @@ Router.prototype._notfoundRoute = function(context) {
   this._invalidateTracker();
 };
 
-Router.prototype.initialize = function() {
+Router.prototype.initialize = function(options) {
+  options = options || {};
+
+  if(this._initialized) {
+    throw new Error("FlowRouter is already initialized");
+  }
+
   var self = this;
   this._updateCallbacks();
 
@@ -312,21 +366,34 @@ Router.prototype.initialize = function() {
   // by overriding page.js`s "show" method.
   // Why?
   // It is impossible to bypass exit triggers,
-  // becuase they execute before the handler and
+  // because they execute before the handler and
   // can not know what the next path is, inside exit trigger.
-  var originalShow = this._page.show;
+  //
+  // we need override both show, replace to make this work
+  // since we use redirect when we are talking about withReplaceState
+  _.each(['show', 'replace'], function(fnName) {
+    var original = self._page[fnName];
+    self._page[fnName] = function(path, state, dispatch, push) {
+      var reload = self.env.reload.get();
+      if (!reload && self._current.path === path) {
+        return;
+      }
 
-  this._page.show = function(path, state, dispatch, push) {
-    var reload = self.env.reload.get();
-    if (!reload && self._current.path === path) {
-      return;
-    }
+      original.call(this, path, state, dispatch, push);
+    };
+  });
 
-    originalShow(path, state, dispatch, push);
-  };
+  // this is very ugly part of pagejs and it does decoding few times
+  // in unpredicatable manner. See #168
+  // this is the default behaviour and we need keep it like that
+  // we are doing a hack. see .path()
+  this._page.base(this._basePath);
+  this._page({
+    decodeURLComponents: true,
+    hashbang: !!options.hashbang
+  });
 
-  // initialize
-  this._page({decodeURLComponents: false});
+  this._initialized = true;
 };
 
 Router.prototype._buildTracker = function() {
@@ -338,8 +405,10 @@ Router.prototype._buildTracker = function() {
       return;
     }
 
-    var route = self._current.route;
-    var path = self._current.path;
+    // see the definition of `this._processingContexts`
+    var currentContext = self._current;
+    var route = currentContext.route;
+    var path = currentContext.path;
 
     if(self.safeToRun === 0) {
       var message =
@@ -355,21 +424,24 @@ Router.prototype._buildTracker = function() {
     // We tackle this with the `safeToRun` variable
     self._globalRoute.clearSubscriptions();
     self.subscriptions.call(self._globalRoute, path);
-    route.callSubscriptions(self._current);
+    route.callSubscriptions(currentContext);
 
     // otherwise, computations inside action will trigger to re-run
     // this computation. which we do not need.
     Tracker.nonreactive(function() {
-      var currentContext = self._current;
-
-      self._processTriggersEnter(currentContext);
-
       var isRouteChange = currentContext.oldRoute !== currentContext.route;
       var isFirstRoute = !currentContext.oldRoute;
       // first route is not a route change
       if(isFirstRoute) {
         isRouteChange = false;
       }
+
+      // Clear oldRouteChain just before calling the action
+      // We still need to get a copy of the oldestRoute first
+      // It's very important to get the oldest route and registerRouteClose() it
+      // See: https://github.com/kadirahq/flow-router/issues/314
+      var oldestRoute = self._oldRouteChain[0];
+      self._oldRouteChain = [];
 
       currentContext.route.registerRouteChange(currentContext, isRouteChange);
       route.callAction(currentContext);
@@ -379,12 +451,14 @@ Router.prototype._buildTracker = function() {
         if(isRouteChange) {
           // We need to trigger that route (definition itself) has changed.
           // So, we need to re-run all the register callbacks to current route
-          // This is pretty important, otherwise tracker 
+          // This is pretty important, otherwise tracker
           // can't identify new route's items
 
           // We also need to afterFlush, otherwise this will re-run
           // helpers on templates which are marked for destroying
-          currentContext.oldRoute.registerRouteClose();
+          if(oldestRoute) {
+            oldestRoute.registerRouteClose();
+          }
         }
       });
     });
@@ -396,100 +470,53 @@ Router.prototype._buildTracker = function() {
 };
 
 Router.prototype._invalidateTracker = function() {
+  var self = this;
   this.safeToRun++;
   this._tracker.invalidate();
-};
+  // After the invalidation we need to flush to make changes imediately
+  // otherwise, we have face some issues context mix-maches and so on.
+  // But there are some cases we can't flush. So we need to ready for that.
 
-Router.prototype._getRegisterTriggersFn = function(triggers) {
-  var fn = function(triggerFns, options) {
-    options = options || {};
-
-    if (options.only && options.except) {
-      var message = "triggers does not support 'only' and 'except' at the same time.";
-      throw new Error(message);
-    }
-
-    _.each(triggerFns, function(fn) {
-      if (typeof fn !== 'function') {
+  // we clearly know, we can't flush inside an autorun
+  // this may leads some issues on flow-routing
+  // we may need to do some warning
+  if(!Tracker.currentComputation) {
+    // Still there are some cases where we can't flush
+    //  eg:- when there is a flush currently
+    // But we've no public API or hacks to get that state
+    // So, this is the only solution
+    try {
+      Tracker.flush();
+    } catch(ex) {
+      // only handling "while flushing" errors
+      if(!/Tracker\.flush while flushing/.test(ex.message)) {
         return;
       }
 
-      if (options.only) {
-        fn._only = {};
-        _.each(options.only, function(name) {
-          fn._only[name] = 1;
-        });
-      }
+      // XXX: fix this with a proper solution by removing subscription mgt.
+      // from the router. Then we don't need to run invalidate using a tracker
 
-      if (options.except) {
-        fn._except = {};
-        _.each(options.except, function(name) {
-          fn._except[name] = 1;
-        });
-      }
+      // this happens when we are trying to invoke a route change
+      // with inside a route chnage. (eg:- Template.onCreated)
+      // Since we use page.js and tracker, we don't have much control
+      // over this process.
+      // only solution is to defer route execution.
 
-      triggers.push(fn);
-    });
-  };
-
-  return fn;
-};
-
-Router.prototype._shouldCallTrigger = function(current, fn) {
-  var name = current.route.name;
-  var shouldCall;
-
-  if (typeof fn !== 'function') {
-    return false;
-  }
-
-  if (fn._only) {
-    shouldCall = !!fn._only[name];
-  } else if (fn._except) {
-    shouldCall = !fn._except[name];
-  } else {
-    shouldCall = true;
-  }
-
-  return shouldCall;
-};
-
-Router.prototype._processTriggersEnter = function(current) {
-  var self = this;
-
-  _.each(this._triggersEnter, function(fn) {
-    if (self._shouldCallTrigger(current, fn)) {
-      fn(current);
-    }
-  });
-};
-
-Router.prototype._processTriggersExit = function(ctx, next) {
-  var self = this;
-
-  _.each(self._triggersExit, function(fn) {
-    if (self._shouldCallTrigger(self._current, fn)) {
-      fn(self._current);
-    }
-  });
-
-  next();
-};
-
-Router.prototype._registerRouteTriggersExit = function(route) {
-  var self = this;
-
-  if (route._triggersExit.length > 0) {
-    // add route's exit triggers
-    self._page.exit(route.path, function(ctx, next) {
-      _.each(route._triggersExit, function(fn) {
-        if (typeof fn === 'function') {
-          fn(self._current);
+      // It's possible to have more than one path want to defer
+      // But, we only need to pick the last one.
+      // self._nextPath = self._current.path;
+      Meteor.defer(function() {
+        var path = self._nextPath;
+        if(!path) {
+          return;
         }
-      });
 
-      next();
-    });
+        delete self._nextPath;
+        self.env.reload.withValue(true, function() {
+          self.go(path);
+        });
+      });
+    }
   }
 };
 
@@ -499,20 +526,60 @@ Router.prototype._updateCallbacks = function () {
   self._page.callbacks = [];
   self._page.exits = [];
 
-  // add global middleware
-  _.each(self._middleware, function(fn) {
-    self._page("*", fn);
-  });
-
   _.each(self._routes, function(route) {
-    self._page(route.path, route._handler);
-    self._registerRouteTriggersExit(route);
+    self._page(route.pathDef, route._actionHandle);
+    self._page.exit(route.pathDef, route._exitHandle);
   });
-
-  self._page.exit("*", self._processTriggersExit.bind(self));
 
   self._page("*", function(context) {
     self._notfoundRoute(context);
+  });
+};
+
+Router.prototype._initTriggersAPI = function() {
+  var self = this;
+  this.triggers = {
+    enter: function(triggers, filter) {
+      triggers = Triggers.applyFilters(triggers, filter);
+      if(triggers.length) {
+        self._triggersEnter = self._triggersEnter.concat(triggers);
+      }
+    },
+
+    exit: function(triggers, filter) {
+      triggers = Triggers.applyFilters(triggers, filter);
+      if(triggers.length) {
+        self._triggersExit = self._triggersExit.concat(triggers);
+      }
+    }
+  };
+};
+
+Router.prototype.wait = function() {
+  if(this._initialized) {
+    throw new Error("can't wait after FlowRouter has been initialized");
+  }
+
+  this._askedToWait = true;
+};
+
+Router.prototype.onRouteRegister = function(cb) {
+  this._onRouteCallbacks.push(cb);
+};
+
+Router.prototype._triggerRouteRegister = function(currentRoute) {
+  // We should only need to send a safe set of fields on the route
+  // object.
+  // This is not to hide what's inside the route object, but to show
+  // these are the public APIs
+  var routePublicApi = _.pick(currentRoute, 'name', 'pathDef', 'path');
+  var omittingOptionFields = [
+    'triggersEnter', 'triggersExit', 'action', 'subscriptions', 'name'
+  ];
+  routePublicApi.options = _.omit(currentRoute.options, omittingOptionFields);
+
+  _.each(this._onRouteCallbacks, function(cb) {
+    cb(routePublicApi);
   });
 };