+++ server/plugins/fast_render.js
@@ -12,8 +12,8 @@ Meteor.startup(function () {
 });
 
 function setupFastRender () {
-  _.each(FlowRouter._routesMap, function (route, path) {
-    FastRender.route(path, function (params, path) {
+  _.each(FlowRouter._routes, function (route) {
+    FastRender.route(route.pathDef, function (routeParams, path) {
       var self = this;
 
       // anyone using Meteor.subscribe for something else?
@@ -25,7 +25,9 @@ function setupFastRender () {
       route._subsMap = {};
       FlowRouter.subscriptions.call(route, path);
       if(route.subscriptions) {
-        route.subscriptions(params);
+        var queryParams = routeParams.query;
+        var params = _.omit(routeParams, 'query');
+        route.subscriptions(params, queryParams);
       }
       _.each(route._subsMap, function (args) {
         self.subscribe.apply(self, args);