+++ test/client/router.subs_ready.spec.js
@@ -12,11 +12,14 @@ Tinytest.addAsync('Client - Router - subsReady - with no args - all subscription
   };
 
   FlowRouter.go('/' + rand);
-  setTimeout(function() {
-    test.isTrue(!!FlowRouter.subsReady());
-    FlowRouter.subscriptions = Function.prototype;
-    next();
-  }, 100);
+
+  Tracker.autorun(function(c) {
+    if(FlowRouter.subsReady()) {
+      FlowRouter.subscriptions = Function.prototype;
+      next();
+      c.stop();
+    }
+  });
 });
 
 Tinytest.addAsync('Client - Router - subsReady - with no args - all subscriptions does not ready', function (test, next) {
@@ -95,11 +98,13 @@ Tinytest.addAsync('Client - Router - subsReady - with args - all subscriptions r
   };
 
   FlowRouter.go('/' + rand);
-  setTimeout(function() {
-    test.isTrue(!!FlowRouter.subsReady('foo', 'baz'));
-    FlowRouter.subscriptions = Function.prototype;
-    next();
-  }, 100);
+  Tracker.autorun(function(c) {
+    if(FlowRouter.subsReady('foo', 'baz')) {
+      FlowRouter.subscriptions = Function.prototype;
+      next();
+      c.stop();
+    }
+  });
 });
 
 Tinytest.addAsync('Client - Router - subsReady - with args - all subscriptions does not ready', function (test, next) {
@@ -215,27 +220,6 @@ Tinytest.addAsync('Client - Router - subsReady - no subscriptions - simple', fun
   FlowRouter.go('/' + rand);
   setTimeout(function() {
     test.isTrue(FlowRouter.subsReady());
-    next();
-  }, 100);
-});
-
-Tinytest.addAsync('Client - Router - ready - deperearted, but still supports', function (test, next) {
-  var rand = Random.id();
-  FlowRouter.route('/' + rand, {
-    subscriptions: function(params) {
-      this.register('bar', Meteor.subscribe('bar'));
-      this.register('foo', Meteor.subscribe('foo'));
-    }
-  });
-
-  FlowRouter.subscriptions = function () {
-    this.register('baz', Meteor.subscribe('baz'));
-  };
-
-  FlowRouter.go('/' + rand);
-  setTimeout(function() {
-    test.isTrue(!!FlowRouter.ready('foo', 'baz'));
-    FlowRouter.subscriptions = Function.prototype;
     next();
   }, 100);
 });