+++ test/common/router.path.spec.js
@@ -1,10 +1,12 @@
 Router = FlowRouter.Router;
 
-Tinytest.add('Common - Router - validate path definition', function (test, next) {
+Tinytest.addAsync('Common - Router - validate path definition', function (test, next) {
   // path must start with '/'
-  test.throws(function() {
+  try {
     FlowRouter.route(Random.id());
-  });
+  } catch(ex) {
+    next();
+  }
 });
 
 Tinytest.add('Common - Router - path - generic', function (test) {
@@ -56,7 +58,7 @@ Tinytest.add('Common - Router - path - missing fields', function (test) {
   var fields = {
     blogId: "1001",
   };
-  var expectedPath = "/blog/1001/some/";
+  var expectedPath = "/blog/1001/some";
 
   var path = FlowRouter.path(pathDef, fields);
   test.equal(path, expectedPath);
@@ -85,7 +87,7 @@ Tinytest.add('Common - Router - path - optional last param missing', function (t
   var fields = {
     blogId: "1001"
   };
-  var expectedPath = "/blog/1001/some/";
+  var expectedPath = "/blog/1001/some";
 
   var path = FlowRouter.path(pathDef, fields);
   test.equal(path, expectedPath);
@@ -115,6 +117,14 @@ Tinytest.add('Common - Router - path - remove trailing slashes', function (test)
   test.equal(path, expectedPath);
 });
 
+Tinytest.add('Common - Router - path - handle multiple slashes', function (test) {
+  var pathDef = "/blog///some/hi////";
+  var expectedPath = "/blog/some/hi";
+
+  var path = FlowRouter.path(pathDef);
+  test.equal(path, expectedPath);
+});
+
 Tinytest.add('Common - Router - path - keep the root slash', function (test) {
   var pathDef = "/";
   var fields = {};
@@ -122,4 +132,4 @@ Tinytest.add('Common - Router - path - keep the root slash', function (test) {
 
   var path = FlowRouter.path(pathDef, fields);
   test.equal(path, expectedPath);
-})
+});